+++ swanlab/api/upload/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+r"""
+@DATE: 2024/4/7 16:56
+@File: __init__.py
+@IDE: pycharm
+@Description:
+    上传相关接口
+"""
+from ..http import get_http
+
+
+
